[java] Moving tests that should belong to client source subtree
