+++ java/server/test/org/openqa/selenium/remote/server/RemoteServerTests.java
@@ -36,13 +36,11 @@ import org.openqa.selenium.remote.server.xdrpc.CrossDomainRpcLoaderTest;
     CrossDomainRpcLoaderTest.class,
     DefaultSessionTest.class,
     DriverFactoryTest.class,
-    NewSessionPayloadTest.class,
     NewSessionPipelineTest.class,
     ResponsesTest.class,
     ResultConfigTest.class,
     SendKeyToActiveElementTest.class,
     SessionLogsTest.class,
-    SyntheticNewSessionPayloadTest.class,
     UploadFileTest.class,
     org.openqa.selenium.remote.server.commandhandler.UploadFileTest.class,
     UtilsTest.class,
